requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
astroid==2.0.4
astunparse==1.6.3
bleach==1.5.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Django==3.0.6
Flask==1.1.2
gast==0.3.3
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==0.9999999
idna==2.9
importlib-metadata==1.6.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.3.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
pandas==1.0.3
Pillow==7.1.2
protobuf==3.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.15.0
sqlparse==0.3.1
tensorboard==1.7.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==1.7.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
typed-ast==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0 
 